Stop supressing exceptions during content indexing
I am sure there could be a subtype of Exception created here, but we do
not use the type of the exception anywhere, and the parent methods also
throw Exception :shrug:

These exceptions prevent indexing from proceeding on error, which has
happened several times recently with fairly undesired consequences.
Hard failures like these could leave the indices in invalid states,
many of which will be dealt with by the
expungeAnyContentTypeIndicesRelatedToVersion check at the start of
buildElasticSearchIndex.