Avoid repeated database calls augmenting groups
The convertGroupsToDTOs used to have a lookup cache for group owners,
but not additional managers. However, the method loaded the manager IDs
and then the user data for each manager _one_ by _one_, group by group;
this could make hundreds of database queries, even if they were no
duplicate users in the lists.

This commit adds a new method for bulk loading additional manager data
for multiple groups at once, then refactors the group conversion to use
the new method. The most efficient approach is to make a unique set of
all owners and managers and load their user data once, so that is what
the code does. Even though the streams approach means we now loop though
the "groups" list ~5 times rather than once, the reduction in database
queries from 2*n_groups to 2 is a massive performance improvement.

The tightening from Iterable to Collection was necessary to use Streams,
and since we only pass in collections anyhow, caused no issues.
Also suppress the warnings about missing users to debug.

The method does try a new approach to the existing "query multiple"
problem; rather than using "... IN (?, ?, ?, ...)", this uses ARRAYs
and does "... = ANY(?)". The query plan is identical, but we have not
used java.sql.Array for selection before so this is an experiment.