Fix anonymisation script column confusion
The "anonymous_user" column does not indicate that the user _is_ an
anonymous user, only that they _were_ anonymous when they performed the
action recorded.
Once the data is anonymised, there is no easy way to distinguish the
user IDs from anon IDs without an expensive JOIN. This column now does
so, but is named to avoid confusion with "anonymous_user" since it is
not the converse of that column!