Replace unnecessary multi-user auth provider query
We never use the multi-user authentication provider query except in
tests, and there we use it with singleton lists anyway. The query was
not well constructed, and a single-user version is much simpler, safer,
and far more readable.

This commit also removes the unused method from the interface, and
alters the tests to use the single-user method.