Swap deprecated Validate.notNull to Objects.requireNonNull
These two methods are identical under the hood (indeed the former just
calls the latter internally). This isn't the recommended fix, which is
to add String error messages in every case, but this is an old code
style and the deprecation warnings are swamping all other warnings in
the codebase!