Stop showing unpublished pages in search results
In a 3VL world, "not published=false" is not the same thing as
"published=true".
If published is not set (i.e. it is NULL) then it is indeed not equal
to false - but it is still not supposed to be published!

Compare this to GitContentManager::getBaseFilters(), where we already
correctly mustMatch a "true" value, which is used when finding content
for other endpoints.