Improve warnings on creating duplicate account
Some user soft-deletion has caused the existing code-path that checks
for duplicate accounts to miss some cases.
This modifies the INSERT query to fail softly, which allows us to
distinguish between database errors and (failed) duplicate inserts. It
is also apparently more performant:
https://aws.amazon.com/blogs/database/hidden-dangers-of-duplicate-key-violations-in-postgresql-and-how-to-avoid-them/

We already had a `if (pst.executeUpdate() == 0)` check, so this just
adjusts the messaging to be consistent with the other pathways.
The error message about being unable to set a password was inaccurate,
so that has been updated too.