Add admin endpoint to upgrade password algorithms
This takes an annoyingly large amount of plumbing through so many
manager layers. It is minimally functional; it only upgrades hashes
that are compatible with the algorithm specified by the request and
does not do anything clever.
The sole reason to have the endpoint at all is just to ensure that
sensitive credentials do not need to be edited by hand or written
elsewhere on the server(s).