Fix manual Orika conversion for InlineRegions
There is no need to do the conversion explicitly like this. The problem
is that "IsaacQuestionBase" is not "ContentBase" and so Orika does not
know to use the clever base converter.
This simply tells Orika to use the clever base converter for
IsaacQuestionBase objects, and we're done!

I feel like maybe there's a better way than making the new class have a
base converter instance inside it, but this was quickest!