Add a page_id column to question_attempts
This prepares the ground for a move to using this column when loading
question attempts from the database. We need to record the value for
new rows before we can start migrating older rows to include the new
column.
Currently the column cannot have a NOT NULL property, but this can be
added later once the data is backfilled.