Remove support for user badges
We haven't made any use of user badges in the 5 or so years they have
been in the codebase; but they have added a performance and maintenance
cost. This PR removes the badges and the overhead, returning log
management directly to the PgLogManager.
It makes no attempt to clean up or remove historic badge data.