Allow page-level types to be deprecated and superseded
This is a change from _any_ content being able to be deprecated, to just
top-level page types (i.e. those that extend SeguePage). We did not
do anything with deprecated set on any other content, so this change is
safe to make and reflects the actual behaviour.
This also moves supersededBy to the SeguePage and subclasses, so that
more than just question pages can be superseded.
It also tidies up other classes that had one or both of these properties
but where we were not using them. A similar change to the frontend types
exists to demonstrate that we only used these for SeguePage level
objects.

I have tested that this has not broken any existing content, and also
verified that ContentSummaryDTOs still get the two properties mapped
across as expected.