Fail hard on ElasticSearch bulk indexing failures
I think this is an improvement, since there are failure modes here where
some items are skipped and that ought to cause the whole index to fail
but did not.
However, it's possible some types of error may not need to be so fatal.
Only one way to find out!