Add missing index to quiz_assignments
We load quizzes set by group on a regular basis, and less often (but
every time a new quiz is created) we load by group and quiz_id. The
ordering of the columns here allows the same index to provide both
as index-only scans.
Just groups would likely be enough too, should this index ever become
large; most groups will not have many quizzes set, so settling for a
linear scan of those would allow quiz_id to be removed from the index.