Avoid double-loading full gameboard each request
We load the gameboard initially not just to check it exists but also to
load the question IDs; these are used to speed up loading question
attempts a lot.
However, the full gameboard requires going to ElasticSearch and then
mapping a bunch of objects between classes. Since we never use anything
from these objects that is not in a Lite (unaugmented) gameboard this
is entirely wasted work.
The end effect is minor since loading question attempts and the full
gameboard still dominate, but I still think worthwhile.