Avoid unnecessarily reregistering Segue Jobs
Currently we reregister jobs on every API start, even if they are
unchanged from the version existing in the database. (The code of the
job can change without the _job_ changing; the job records the class to
execute and the cron record).
Doing so overwrites the last run time of the job, which prevents us from
doing sensible things if the API has been offline for a while.

This change only reregisters the job if the job details or cron string
change.