Sort users in groups more consistently with Excel
This isn't the _right_ fix, which would involve using collations
correctly and working out what language we wanted to sort using. But
assuming most people are using Excel in English, this fixes the most
common case where sort does not match; apostrophes in names.
Just removing the character does not lead to a consistent sort, however,
so it may be worth reviewing that if people notice.

This goes back to the in-place approach from the original PR, since
comparator chaining like this is nicer.

Neaten the tests, add more test cases for the apostrophe cases.