Allow sending and saving of account deletion tokens
This adds a new table for account deletion tokens, a new persistence
manager, an endpoint to request the token via, and adds assorted other
methods for generating secure random tokens and creating the emails.

There are still questions about whether the endpoint is in the right
place, and whether the methods should live elsewhere in the tree of
managers.