Store last seen time in a separate table to users
Postgres has an MVCC policy that causes the entire row to be copied if
it is updated. Most details of a users table row are relatively static,
but we update "last_seen" on a ~5 minutely basis when a user is active.
This causes the whole row to be copied, which bloats the index and the
table.

However whilst these changes do remove that bloat and performance
impact, this is hard to measure. We also query for users orders of
magnitude more frequently than we update last seen.
The need for a join for every query to load a user slows down loading of
a single user by ~40% in my tests, and slows down loading a group of
users by almost 100%. It is not clear that this is worth the gain.