Only clean up ES indices when single indexer running
If more than one indexing process is happening, cleaning up unaliased
indices is not a safe operation. There are several different ways the
threads can conflict!
Much of the problem is caused by the fact each "version" is a collection
of 5 indices {content, content_error, metadata, published_unit, unit}
which each must get an alias. Each of these indices is created, and then
aliased, separately.

1) Job A is in the process of creating a new version's collection of
   ES indices; none of them have aliases yet. Job B concurrently reaches
   the cleanup stage, finds the indices Job A is adding data to, sees
   they are unaliased and deletes them. This causes Job A to fail, but
   this failure should be noticed and fail safe(ish).

2) Job A has created all of the collection of indices for a new version,
   and is starting to move the various alias pointers. Job B reaches the
   cleanup stage, finds some mixture of two versions' indices without
   aliases and deletes them. This may cause Job A to fail, but will also
   leave the not-yet-aliased-as previous version in an invalid state;
   this version is being used by the API and the live site is broken.

This code now uses a thread-safe counter to skip running the deletion if
another indexing job is running. However, this operation is not itself
thread safe; it may take some time between the check and the response
from ElasticSearch listing the indices and their aliases, in which time
a new indexing job may have started creating indices. So we check again
after getting the index list request that there are still no other
concurrent jobs. This should catch all plausible cases.

Without truly serialising the whole ETL mechanics, I do not think we can
make it perfect - but I believe this is a net improvement.