Extend CoordinateItems and Questions to n dimensions
Instead of a hardcoded x and y value, now use a list of Strings to
represent the coordinates. This also requires the placeholders to be a
list of values too.

In order not to break existing questions whilst the changes are
deployed, we leave the existing x, y values in the model. Then the
content can be updated with both, and we can remove the deprecated
values in a later change.