Allow coordinate questions to have n dimensions
This moves the validator to use the new list-of-strings representation
of coordinates to allow an arbitrary number of dimensions in a
coordinate.
There are some subtleties, because "ordering" the coordinates for
comparison is complicated - as is the quadratic nature of the checking
algorithm.

It also cleans up some log messages and comments for clarity.