Only use sig figs for coord question marking if set
We were defaulting to 2sf if the sig figs were not set for coordinate
questions. Adding feedback about sig figs made the content team realise
this and that it was not what was expected.
Now, only validate using sig figs if they are provided in the editor.
Also, warn if inconsistent values are used for these bounds.