Allow shuffling of question finder results
This does not use the existing but deprecated GitContentManager method,
but updates the questionSearch method to allow passing a random seed
which will be used if no sort order is provided. The random sorting is
done in ElasticSearch in exactly the same way the old gameboard builder
randomisation was.

I reordered the arguments of the function to better logically group
them, and removed some unused arguments from all the methods involved.