Stop special-casing tutors for quizzes
In hiddenFromRoles, the list should be exhaustive and definitive; if a
role is present, the quiz cannot be seen by them. If a role is not
present, they can see the quiz.
Special-casing tutors was confusing. Now if tutors behave like all other
account types for quiz filtering.

This does require a content change to make consistent with the old
behaviour, but that change has now been done.