Expose DO to DTO mapping via GitContentManager
Previously, if you needed a DO and a DTO for something (such as
augmenting a page), either the class needed a GitContentManager _and_ a
ContentMapper, or you used both getContentById and getContentDOById.
The latter had the potential for a cache-miss and an extra round-trip
to ElasticSearch. But all getContentById did was call getContentDOById
and then map it; so this just exposes that mapping on the GCM.
Given the public access in the ContentMapper, I don't think there is any
additional risk of mis-using the method here.

Conveniently there were several places that used this functionality
that can be moved to using this new method, which will also be useful
for proposed new endpoints.
Also clean up the unused ContentMapper in the ContactFacade.