Improve gameboard loading for bulk queries
At present when gameboards are loaded in bulk and augmented by
getGameboardsByIds in the GameboardPersistenceManager, we make separate
requests to ElasticSearch for each gameboard. This is because the DO to
DTO mapping is run once per board, rather than all together, and each
call makes a query to ES.
But getGameboardItemMap, which that method implicitly used, had been
written to allow bulk loading of data in much larger groups than a
single gameboard. So instead of making the bulk convertToGameboardDTOs
method use the single convertToGameboardDTO method, the reverse is now
true; single boards are augmented via a singleton list and the bulk
loading is the only code path.
That enables DO to DTO mapping to vastly reduce the number of ES queries
that are made, and also reduced the spaghetti code down to a single code
path. From there, all the augmentation methods can be combined and the
duplicate or redundant code removed.

Doing this also allows us to fix a bug where existing bulk augmentation
of gameboards would lose the creationContext for a question if it was
included on multiple boards. This was because there was one map of
question ID -> gameboard item (gameboardReadyQuestions), which could
only keep just one context per question.
This rewrites the code to use a map of question ID to ContentDTO (which
is always the same), and does the mapping inline in the renamed
augmentGameboardItemsWithContentData method using the correct context
for each question/gameboard combination.