Ensure question randomisation is purely random
Previously we were combining the question's ranking by how closely it
matched the provided filters with the random ranking. The combination
function defaults to multiplication, which meant that questions with
high ranks remained high when multiplied by a random value in (0, 1).
This meant some questions remained prominent on shuffling.
By replacing the rank score with just a random number, results are now
entirely random in order.