Improve logged event detail for question finder
In order to reason about what users are doing, and safely deprecate the
older parameters, it would be best if we logged everything.
This adds the missing params, but also removes more empty or unnecessary
values from the stored log value by using the fact that null values will
not be serialised.