Add an endpoint to load a detailed quiz summary
Currently there is no way to see the rubric of a quiz (to work out if it
is relevant to you, say) before freely attempting it. However, if you
freely attempt it, you must complete it to remove it from your account.

This endpoint allows previewing the rubric for non-teacher users for all
freely-attemptable quizzes.