Only augment related questions with attempt data
Previously all related content types were augmented with "NOT_ATTEMPTED"
if they were not question pages. This makes no sense for quizzes or
general pages which cannot have question attempts stored.

However, some concept pages had questions on and those were being
augmented and this will break that. Since questions on concept pages are
not supported (because we hide irrelevant sections from users, what
would it mean to "complete" a concept page?) this is okay for now.

This neatens up places relying on fasttrack and ordinary question page
types, too, to a single new constant in list and set form.