Move some quiz tests to integration testing
Also fix a bug in the existing integration tests where failure to create
an assignment was missed by the test.

These unit tests were blocking addressing performance issues on loading
a users' quiz assignments (because the mocking was too complex to
update for the changes).

The methods:
createQuizAssignmentEndpoint_assignDuplicateQuizAsTeacherAfterFirstDue_succeeds
createQuizAssignmentEndpoint_assignDuplicateQuizAsTeacher_fails
both use direct access to the database manager to initialise the test
conditions. That may not be a sensible approach.