Synchronise users with external service more slowly
As an attempt to hit the rate limit only after more successful requests,
add a small delay between users in the loop.
We retry the job after 4 hours, so it is unlikely that two will end up
concurrent, but mark the job as needing to be unique.

The next thing to try will be to run the job every 2 hours, rather than
every 4 hours.