Further warning logs on request IP extraction
If the API is directly exposed to the client, we may not trust the
headers used to find the client IP address in this class. This is
unlikely, since the current Jetty configuration offers no TLS or any
meaningful security headers.

To future-proof things, and to protect anyone naively re-using the code,
add some log messages triggered by the two types of unexpected
behaviour.

If a request comes in without these headers, this may be local
development, so only log a warning once. But if subsequent requests
after a direct request set proxy headers we then use, log this as a
warning since this is likely malicious behaviour.