Clarify streams code augmenting gameboards
This does not change the behaviour right now because it was not a
parallel stream, but the "unordered" could have reduced the likelihood
of cache hits, and reduced the efficiency of the code.
The order of the list seems stable between requests, so additional
sorting does not seem necessary to ensure cache use.