use non-static cache
Playing around with the server a little bit, I see lots of instances of
MicrosoftAuthenticator being created (and, interestingly, each log-in
request creates a new instance), but the instance being used is always
the same: the one that's been passed in to userAuthenticationManager,
which is a singleton. It's possible that my local app is uses less
threads than the live app, but even on the live app. However, because
the exchange and token access happen while serving the same request,
I think it's guaranteed that we only ever access the cache from within
the same thread.