Batch bulk requests to ElasticSearch
Locally, the bulk index requests were erroring after overloading the
internal ElasticSearch indexer queue. We were sending requests with
~70,000 JSON documents in, and this was too much for the server to
process in one go.
I tested partitions of 1000, 5000, 10000 and 100000. The larger the
partition, the faster the indexing, except the 100000 case which just
errors in the same way the non-batched request does for the same reason.