Make board wildcards optional and default empty
We have not shown wildcards on ordinary boards in a long time, so this
moves to not _adding_ them to ordinary boards either. No more random
selection of a wildcard; the wildcard is just left empty if it is not
explicitly provided by a staff user. All other users get no wildcard
when they make a board now.