Always cache ResultsWrappers and separate DO/DTOs
Rather than using Optionals, instead cache a ResultsWrapper object; this
can just contain an empty list in the Not Found case, and a single
result in the successful getById case.
To avoid unsafe casts, use separate caches for DOs and DTOs; this is
perhaps unnecessary but does make the Java a lot cleaner.

This also moves the DTO method to using the idiomatic and atomic get()
method of the cache, which is again cleaner. By using this method, we
should actually get usable stats from the stats collection on the
caches; previously we never saw a miss because we always checked before
trying to get something from the cache.