Compute quiz feedback outside of student loop
This does not need to be recomputed once per student in the group,
because if it is we end up with n*n*m queries for n students and m quiz
assignments. This way ends up doing 2*n*m queries, which whilst terrible
is still a vast improvement for large groups.