Read marksAwarded value for LLMFreeText responses
In the context of the markbook/assignment progress, LLMFreeTextQuestionValidationResponses have a marksAwarded field that we need to extract the full question attempt to read, but we want to keep these attempts lightweight for all other question parts to minimise processing unnecessary data.

This change checks each question part and extracts the full response for only LLMFreeText ones.