Add quiz_attempts index migration
When loading free quiz attempts, we scan the quiz_attempts table to find
quizzes by user_id. We also query by user_id and quiz_id in a few
endpoints. We currently never load quiz_attempts by just quiz_id, since
there is little need to know what users have done a specific quiz.

The existing index was on (quiz_id, user_id), which worked well for the
paired loading, but could not be used to reduce the search space when
loading by user_id. This replacement index still works for paired
loading, but can also be used to reduce the search space when loading
by user_id.

Has already been run on live, as part of the testing that it worked.